UPDATE print patch and refactor
